refactor: move upgrade tab function to lib
