+++ apps/meteor/client/views/hooks/useUpgradeTabParams.ts
@@ -2,7 +2,7 @@ import { useSetting, useEndpoint } from '@rocket.chat/ui-contexts';
 import { useQuery } from '@tanstack/react-query';
 import { format } from 'date-fns';
 
-import { UpgradeTabVariant, getUpgradeTabType } from '../../../lib/getUpgradeTabType';
+import { getUpgradeTabType, UpgradeTabVariant } from '../../../lib/upgradeTab';
 
 export const useUpgradeTabParams = (): { tabType: UpgradeTabVariant | false; trialEndDate: string | undefined; isLoading: boolean } => {
 	const getRegistrationStatus = useEndpoint('GET', '/v1/cloud.registrationStatus');
