+++ apps/meteor/tests/unit/lib/upgradeTab.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import { describe, it } from 'mocha';
 
-import { getUpgradeTabType } from '../../../lib/getUpgradeTabType';
+import { getUpgradeTabLabel, getUpgradeTabType, isFullyFeature } from '../../../lib/upgradeTab';
 
 describe('getUpgradeTabType()', () => {
 	it("should return 'go-fully-featured'", () => {
@@ -74,5 +74,33 @@ describe('getUpgradeTabType()', () => {
 				hasGoldLicense: false,
 			}),
 		).to.be.equal(false);
+	});
+});
+
+describe('getUpgradeTabLabel()', () => {
+	it("should return 'Upgrade_tab_go_fully_featured'", () => {
+		expect(getUpgradeTabLabel('go-fully-featured')).to.be.equal('Upgrade_tab_go_fully_featured');
+		expect(getUpgradeTabLabel('go-fully-featured-registered')).to.be.equal('Upgrade_tab_go_fully_featured');
+	});
+
+	it("should return 'Upgrade_tab_trial_guide'", () => {
+		expect(getUpgradeTabLabel('trial-gold')).to.be.equal('Upgrade_tab_trial_guide');
+		expect(getUpgradeTabLabel('trial-enterprise')).to.be.equal('Upgrade_tab_trial_guide');
+	});
+
+	it("should return 'Upgrade_tab_upgrade_your_plan'", () => {
+		expect(getUpgradeTabLabel('upgrade-your-plan')).to.be.equal('Upgrade_tab_upgrade_your_plan');
+		expect(getUpgradeTabLabel(false)).to.be.equal('Upgrade_tab_upgrade_your_plan');
+	});
+});
+
+describe('isFullyFeature()', () => {
+	it("should return 'true", () => {
+		expect(isFullyFeature('go-fully-featured')).to.be.equal(true);
+	});
+
+	it("should return 'false", () => {
+		expect(isFullyFeature('upgrade-your-plan')).to.be.equal(false);
+		expect(isFullyFeature(false)).to.be.equal(false);
 	});
 });
